+++ frontend/catalyst/jax_tracer.py
@@ -15,6 +15,7 @@
 """
 # pylint: disable=too-many-lines
 
+import logging
 from dataclasses import dataclass
 from functools import partial, reduce
 from typing import Any, Callable, Dict, List, Optional, Sequence, Tuple, Union
@@ -78,12 +79,16 @@ from catalyst.jax_primitives import (
     tensorobs_p,
     var_p,
 )
+from catalyst.logging import debug_logger, debug_logger_init
 from catalyst.tracing.contexts import (
     EvaluationContext,
     EvaluationMode,
     JaxTracingContext,
 )
 from catalyst.utils.exceptions import CompileError
+
+logger = logging.getLogger(__name__)
+logger.addHandler(logging.NullHandler())
 
 
 class Function:
@@ -99,10 +104,12 @@ class Function:
         AssertionError: Invalid function type.
     """
 
+    @debug_logger_init
     def __init__(self, fn):
         self.fn = fn
         self.__name__ = fn.__name__
 
+    @debug_logger
     def __call__(self, *args, **kwargs):
         jaxpr, out_tree = make_jaxpr2(self.fn)(*args)
 
@@ -129,6 +136,7 @@ PAULI_NAMED_MAP = {
 }
 
 
+@debug_logger
 def retrace_with_result_types(jaxpr: ClosedJaxpr, target_types: List[ShapedArray]) -> ClosedJaxpr:
     """Return a JAXPR that is identical to the given one but with added type conversion operations
     to produce the provided type signature in its output."""
@@ -153,6 +161,7 @@ def retrace_with_result_types(jaxpr: ClosedJaxpr, target_types: List[ShapedArray
     return ClosedJaxpr(jaxpr2, consts)
 
 
+@debug_logger
 def unify_jaxpr_result_types(jaxprs: List[ClosedJaxpr]) -> List[ClosedJaxpr]:
     """Unify result signatures across a set of JAXPRs by promoting to common types.
 
@@ -188,10 +197,12 @@ class QRegPromise:
     """QReg adaptor tracing the qubit extractions and insertions. The adaptor works by postponing
     the insertions in order to re-use qubits later thus skipping the extractions."""
 
+    @debug_logger_init
     def __init__(self, qreg: DynamicJaxprTracer):
         self.base: DynamicJaxprTracer = qreg
         self.cache: Dict[Any, DynamicJaxprTracer] = {}
 
+    @debug_logger
     def extract(self, wires: List[Any], allow_reuse=False) -> List[DynamicJaxprTracer]:
         """Extract qubits from the wrapped quantum register or get the already extracted qubits
         from cache"""
@@ -215,6 +226,7 @@ class QRegPromise:
                 qubits.append(qextract_p.bind(qrp.base, w))
         return qubits
 
+    @debug_logger
     def insert(self, wires, qubits) -> None:
         """Insert qubits to the cache."""
         qrp = self
@@ -225,6 +237,7 @@ class QRegPromise:
             ), f"Attempting to insert an already-inserted wire {w} into {qrp.base}"
             qrp.cache[w] = qubit
 
+    @debug_logger
     def actualize(self) -> DynamicJaxprTracer:
         """Prune the qubit cache by performing the postponed insertions."""
         qrp = self
@@ -280,6 +293,7 @@ class HybridOp(Operation):
     num_wires = AnyWires
     binder: Callable = _no_binder
 
+    @debug_logger_init
     def __init__(self, in_classical_tracers, out_classical_tracers, regions: List[HybridOpRegion]):
         self.in_classical_tracers = in_classical_tracers
         self.out_classical_tracers = out_classical_tracers
@@ -291,6 +305,7 @@ class HybridOp(Operation):
         nested_ops = [r.quantum_tape.operations for r in self.regions if r.quantum_tape]
         return f"{self.name}(tapes={nested_ops})"
 
+    @debug_logger
     def bind_overwrite_classical_tracers(
         self, ctx: JaxTracingContext, trace: DynamicJaxprTrace, *args, **kwargs
     ) -> DynamicJaxprTracer:
@@ -309,6 +324,7 @@ class HybridOp(Operation):
             eqn.outvars[i] = trace.getvar(t)
         return out_quantum_tracer
 
+    @debug_logger
     def trace_quantum(
         self,
         ctx: JaxTracingContext,
@@ -329,6 +345,7 @@ def has_nested_tapes(op: Operation) -> bool:
     )
 
 
+@debug_logger
 def trace_to_jaxpr(func, static_argnums, abstracted_axes, args, kwargs):
     """Trace a Python function to JAXPR.
 
@@ -356,6 +373,7 @@ def trace_to_jaxpr(func, static_argnums, abstracted_axes, args, kwargs):
     return jaxpr, out_treedef
 
 
+@debug_logger
 def lower_jaxpr_to_mlir(jaxpr, func_name):
     """Lower a JAXPR to MLIR.
 
@@ -384,6 +402,7 @@ def lower_jaxpr_to_mlir(jaxpr, func_name):
     return mlir_module, ctx
 
 
+@debug_logger
 def trace_quantum_tape(
     quantum_tape: QuantumTape,
     device: QubitDevice,
@@ -473,6 +492,7 @@ def trace_quantum_tape(
     return qrp
 
 
+@debug_logger
 def trace_observables(
     obs: Operation, qrp: QRegPromise, m_wires: int
 ) -> Tuple[List[DynamicJaxprTracer], Optional[int]]:
@@ -523,6 +543,7 @@ def trace_observables(
     return obs_tracers, (len(qubits) if qubits else None)
 
 
+@debug_logger
 def pauli_sentence_to_hamiltonian_obs(paulis, qrp: QRegPromise) -> List[DynamicJaxprTracer]:
     """Convert a :class:`pennylane.pauli.PauliSentence` into a Hamiltonian.
 
@@ -544,6 +565,7 @@ def pauli_sentence_to_hamiltonian_obs(paulis, qrp: QRegPromise) -> List[DynamicJ
     return hamiltonian_p.bind(coeffs, *nested_obs)
 
 
+@debug_logger
 def pauli_word_to_tensor_obs(obs, qrp: QRegPromise) -> List[DynamicJaxprTracer]:
     """Convert a :class:`pennylane.pauli.PauliWord` into a Named or Tensor observable.
 
@@ -573,6 +595,7 @@ def identity_qnode_transform(tape: QuantumTape) -> (Sequence[QuantumTape], Calla
 
 
 # pylint: disable=too-many-statements,too-many-branches
+@debug_logger
 def trace_quantum_measurements(
     device: QubitDevice,
     qrp: QRegPromise,
@@ -613,7 +636,10 @@ def trace_quantum_measurements(
 
             if o.return_type.value == "sample":
                 if shots is None:
-                    raise ValueError("qml.sample cannot work with shots=None")
+                    raise ValueError(
+                        "qml.sample cannot work with shots=None. "
+                        "Please specify a finite number of shots."
+                    )
                 if o.mv is not None:  # qml.sample(m)
                     out_classical_tracers.append(o.mv)
                 else:
@@ -632,7 +658,10 @@ def trace_quantum_measurements(
                 out_classical_tracers.append(probs_p.bind(obs_tracers, shape=shape))
             elif o.return_type.value == "counts":
                 if shots is None:
-                    raise ValueError("qml.sample cannot work with shots=None")
+                    raise ValueError(
+                        "qml.sample cannot work with shots=None. "
+                        "Please specify a finite number of shots."
+                    )
                 shape = (2**nqubits,) if using_compbasis else (2,)
                 results = counts_p.bind(obs_tracers, shots=shots, shape=shape)
                 if using_compbasis:
@@ -666,6 +695,7 @@ def trace_quantum_measurements(
     return out_classical_tracers, out_tree
 
 
+@debug_logger
 def is_transform_valid_for_batch_transforms(tape, flat_results):
     """Not all transforms are valid for batch transforms.
     Batch transforms will increase the number of tapes from 1 to N.
@@ -709,6 +739,7 @@ def is_transform_valid_for_batch_transforms(tape, flat_results):
     return are_batch_transforms_valid
 
 
+@debug_logger
 def apply_transform(qnode_program, device_program, device_modify_measurements, tape, flat_results):
     """Apply transform."""
     # Some transforms use trainability as a basis for transforming.
@@ -735,6 +766,7 @@ def apply_transform(qnode_program, device_program, device_modify_measurements, t
     return tapes, post_processing
 
 
+@debug_logger
 def split_tracers_and_measurements(flat_values):
     """Return classical tracers and measurements"""
     classical = []
@@ -752,6 +784,7 @@ def split_tracers_and_measurements(flat_values):
     return classical, measurements
 
 
+@debug_logger
 def trace_post_processing(ctx, trace, post_processing: Callable, pp_args):
     """Trace post processing function.
 
@@ -785,6 +818,7 @@ def trace_post_processing(ctx, trace, post_processing: Callable, pp_args):
         return closed_jaxpr, out_type, out_tree_promise()
 
 
+@debug_logger
 def reset_qubit(qreg_in, w):
     """Perform a qubit reset on a single wire. Suitable for use during late-stage tracing,
     as JAX primitives are used directly. These operations will not appear on tape."""
@@ -810,6 +844,7 @@ def reset_qubit(qreg_in, w):
     return qreg_out
 
 
+@debug_logger
 def trace_quantum_function(
     f: Callable, device: QubitDevice, args, kwargs, qnode=None
 ) -> Tuple[ClosedJaxpr, Any]: